Menú popup terminado e implementada la vista detalles del rocket + link wikipedia

+++ app/src/main/java/com/example/icb0007_uf1_pr01_samuelmateostovar/data/local/RocketEntity.kt
@@ -3,6 +3,7 @@ package com.example.icb0007_uf1_pr01_samuelmateostovar.data.local
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.example.icb0007_uf1_pr01_samuelmateostovar.data.remote.Dimension
+import com.google.gson.annotations.SerializedName
 
 @Entity
 data class RocketEntity(
@@ -11,7 +12,7 @@ data class RocketEntity(
     val name: String,
     val type: String,
     val active: Boolean,
-    val costPerLaunch: Int,
+    val costPerLaunch: Long,
     val successRatePct: Int,
     val country: String,
     val company: String,
